requirements.txt
pyyaml
pandas
duckdb
ollama
streamlit
rich
typer
pytest
python-evtx
sqlparse
